+++ DeltaEngine/Engine/Rendering/Sprite.cpp
@@ -48,10 +48,10 @@ void Sprite::Render(Rendering::Renderer* renderer, const ViewportData& viewportD
 	else
 	{
 		Rendering::Rect srcRect;
-		srcRect.x = _spriteData->spriteStart.GetX();
-		srcRect.y = _spriteData->spriteStart.GetY();
-		srcRect.w = _spriteData->spriteEnd.GetX() - _spriteData->spriteStart.GetX();
-		srcRect.h = _spriteData->spriteEnd.GetY() - _spriteData->spriteStart.GetY();
+		srcRect.x = static_cast<int>(_spriteData->spriteStart.GetX());
+		srcRect.y = static_cast<int>(_spriteData->spriteStart.GetY());
+		srcRect.w = static_cast<int>(_spriteData->spriteEnd.GetX() - _spriteData->spriteStart.GetX());
+		srcRect.h = static_cast<int>(_spriteData->spriteEnd.GetY() - _spriteData->spriteStart.GetY());
 
 		Rendering::RenderCopyEx(renderer, _spriteData->texture, &srcRect, &destRect,
 								transform.rotation, 0,
